# app/web/openapi.json
{
  "components": {
    "schemas": {
      "ActivationCurve": {
        "description": "One cell's response across the sweep's amplitude grid, aligned index-for-index\nwith ``AmplitudeSweepResponse.amplitudes_uA``.\n\nNote there is deliberately no \"activation fraction\" here. The patch is a target\nplus its bystanders (a handful of cells), so a fraction would be a two- or\nthree-level step function wearing the costume of a sigmoid. Per-cell traces are\nwhat the engine actually knows.",
        "properties": {
          "activated": {
            "items": {
              "type": "boolean"
            },
            "title": "Activated",
            "type": "array"
          },
          "blocks": {
            "default": false,
            "title": "Blocks",
            "type": "boolean"
          },
          "cell_id": {
            "title": "Cell Id",
            "type": "string"
          },
          "crossing_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crossing Ua"
          },
          "initiation_region": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Initiation Region",
            "type": "array"
          },
          "is_target": {
            "title": "Is Target",
            "type": "boolean"
          }
        },
        "required": [
          "cell_id",
          "is_target",
          "activated",
          "initiation_region"
        ],
        "title": "ActivationCurve",
        "type": "object"
      },
      "AmplitudeSweepResponse": {
        "properties": {
          "amplitudes_uA": {
            "items": {
              "type": "number"
            },
            "title": "Amplitudes Ua",
            "type": "array"
          },
          "curves": {
            "items": {
              "$ref": "#/components/schemas/ActivationCurve"
            },
            "title": "Curves",
            "type": "array"
          }
        },
        "required": [
          "amplitudes_uA",
          "curves"
        ],
        "title": "AmplitudeSweepResponse",
        "type": "object"
      },
...
      "CadBodySpec": {
        "description": "An imported STEP/BREP solid, referenced by the id ``POST /cad`` returned. The\ngmsh load lives in the FEM env, so the id resolves to a file there, not here.",
        "properties": {
          "conductive_faces": {
            "default": "all",
            "enum": [
              "tip",
              "sides",
              "all"
            ],
            "title": "Conductive Faces",
            "type": "string"
          },
          "kind": {
            "const": "cad",
            "default": "cad",
            "title": "Kind",
            "type": "string"
          },
          "upload_id": {
            "title": "Upload Id",
            "type": "string"
          }
        },
        "required": [
          "upload_id"
        ],
        "title": "CadBodySpec",
        "type": "object"
      },
...
      "CylinderBody": {
        "properties": {
          "conductive_faces": {
            "default": "all",
            "enum": [
              "tip",
              "sides",
              "all"
            ],
            "title": "Conductive Faces",
            "type": "string"
          },
          "height_um": {
            "default": 30.0,
            "exclusiveMinimum": 0.0,
            "title": "Height Um",
            "type": "number"
          },
          "kind": {
            "const": "cylinder",
            "default": "cylinder",
            "title": "Kind",
            "type": "string"
          },
          "radius_um": {
            "default": 5.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius Um",
            "type": "number"
          }
        },
        "title": "CylinderBody",
        "type": "object"
      },
...
      "FieldGridResponse": {
        "description": "Ve (mV) on a square grid at the cell plane. Carries either the synchronous\nanalytical preview or the FEM field solved by ``POST /field/accurate``.",
        "properties": {
          "ve_mV": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Ve Mv",
            "type": "array"
          },
          "vmax_mV": {
            "title": "Vmax Mv",
            "type": "number"
          },
          "xs_um": {
            "items": {
              "type": "number"
            },
            "title": "Xs Um",
            "type": "array"
          },
          "ys_um": {
            "items": {
              "type": "number"
            },
            "title": "Ys Um",
            "type": "array"
          }
        },
        "required": [
          "xs_um",
          "ys_um",
          "ve_mV",
          "vmax_mV"
        ],
        "title": "FieldGridResponse",
        "type": "object"
      },
      "FrustumBody": {
        "properties": {
          "base_radius_um": {
            "default": 8.0,
            "exclusiveMinimum": 0.0,
            "title": "Base Radius Um",
            "type": "number"
          },
          "conductive_faces": {
            "default": "all",
            "enum": [
              "tip",
              "sides",
              "all"
            ],
            "title": "Conductive Faces",
            "type": "string"
          },
          "height_um": {
            "default": 20.0,
            "exclusiveMinimum": 0.0,
            "title": "Height Um",
            "type": "number"
          },
          "kind": {
            "const": "frustum",
            "default": "frustum",
            "title": "Kind",
            "type": "string"
          },
          "top_radius_um": {
            "default": 2.0,
            "minimum": 0.0,
            "title": "Top Radius Um",
            "type": "number"
          }
        },
        "title": "FrustumBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HemisphereBody": {
        "properties": {
          "kind": {
            "const": "hemisphere",
            "default": "hemisphere",
            "title": "Kind",
            "type": "string"
          },
          "radius_um": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius Um",
            "type": "number"
          }
        },
        "title": "HemisphereBody",
        "type": "object"
      },
      "JobStatus": {
        "description": "A background job's state, polled by the client. When ``status`` is ``\"done\"``,\nthe matching result is present: ``scorecard`` for a score job, ``field`` (+\n``max_divergence_pct`` vs the analytical preview) for an accurate-field job,\n``study`` for a geometry sweep, ``sweep`` for an amplitude sweep. ``cached`` means\nit was served from a prior identical run (P7 S3).",
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FieldGridResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "fraction": {
            "title": "Fraction",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_divergence_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Divergence Pct"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "scorecard": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScorecardResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "running",
              "done",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "study": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudyResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "sweep": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AmplitudeSweepResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "fraction",
          "message"
        ],
        "title": "JobStatus",
        "type": "object"
      },
...
      "NoBody": {
        "description": "A flat 2D face on the array plane: the default, analytical-friendly.",
        "properties": {
          "kind": {
            "const": "none",
            "default": "none",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "NoBody",
        "type": "object"
      },
      "SceneControls": {
        "description": "The Compare screen's inputs: the same handful of controls the Dash app\nexposes, translated to specs server-side via ``app.scene.build_scene``.",
        "properties": {
          "body": {
            "discriminator": {
              "mapping": {
                "cad": "#/components/schemas/CadBodySpec",
                "cylinder": "#/components/schemas/CylinderBody",
                "frustum": "#/components/schemas/FrustumBody",
                "hemisphere": "#/components/schemas/HemisphereBody",
                "none": "#/components/schemas/NoBody"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NoBody"
              },
              {
                "$ref": "#/components/schemas/HemisphereBody"
              },
              {
                "$ref": "#/components/schemas/CylinderBody"
              },
              {
                "$ref": "#/components/schemas/FrustumBody"
              },
              {
                "$ref": "#/components/schemas/CadBodySpec"
              }
            ],
            "title": "Body"
          },
          "electrode_um": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Electrode Um",
            "type": "number"
          },
          "extent_um": {
            "default": 130.0,
            "exclusiveMinimum": 0.0,
            "title": "Extent Um",
            "type": "number"
          },
          "include_scorecard": {
            "default": false,
            "title": "Include Scorecard",
            "type": "boolean"
          },
          "layout": {
            "default": "single",
            "enum": [
              "single",
              "bipolar"
            ],
            "title": "Layout",
            "type": "string"
          },
          "n": {
            "default": 61,
            "maximum": 257.0,
            "minimum": 9.0,
            "title": "N",
            "type": "integer"
          },
          "neighbor_um": {
            "default": 40.0,
            "exclusiveMinimum": 0.0,
            "title": "Neighbor Um",
            "type": "number"
          },
          "overlap_policy": {
            "default": "reject",
            "enum": [
              "reject",
              "displace"
            ],
            "title": "Overlap Policy",
            "type": "string"
          },
          "phase_width_us": {
            "default": 200.0,
            "exclusiveMinimum": 0.0,
            "title": "Phase Width Us",
            "type": "number"
          },
          "pitch_um": {
            "default": 60.0,
            "exclusiveMinimum": 0.0,
            "title": "Pitch Um",
            "type": "number"
          },
          "sigma_S_per_m": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Sigma S Per M",
            "type": "number"
          }
        },
        "title": "SceneControls",
        "type": "object"
      },
      "ScorecardResponse": {
        "description": "The evaluator's operating-window verdict. ``activated=False`` means the target\nnever fired in the searched range; only ``offtarget_hash`` comes back with it, and\nevery other field is absent.",
        "properties": {
          "activated": {
            "title": "Activated",
            "type": "boolean"
          },
          "limiting": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limiting"
          },
          "limiting_off_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limiting Off Id"
          },
          "off_min_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Off Min Ua"
          },
          "off_target_thresholds_uA": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Off Target Thresholds Ua"
          },
          "offtarget_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offtarget Hash"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "safe_at_target": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe At Target"
          },
          "safety_ceiling_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Ceiling Ua"
          },
          "target_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Ua"
          },
          "usable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable"
          },
          "usable_margin_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable Margin Ua"
          },
          "window_hi_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Hi Ua"
          },
          "window_lo_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Lo Ua"
          }
        },
        "required": [
          "activated"
        ],
        "title": "ScorecardResponse",
        "type": "object"
      },
      "StudyControls": {
        "description": "A geometry sweep: the diameter \u00d7 pitch grid to explore, plus the patch it is\nscored on. Combos with ``pitch < diameter`` (which would overlap) are dropped.",
        "properties": {
          "aperture_um": {
            "default": 120.0,
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Aperture Um",
            "type": "number"
          },
          "arrangement": {
            "default": "hex",
            "enum": [
              "grid",
              "hex"
            ],
            "title": "Arrangement",
            "type": "string"
          },
          "diameters_um": {
            "items": {
              "type": "number"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Diameters Um",
            "type": "array"
          },
          "neighbor_um": {
            "default": 40.0,
            "exclusiveMinimum": 0.0,
            "title": "Neighbor Um",
            "type": "number"
          },
          "phase_width_us": {
            "default": 200.0,
            "exclusiveMinimum": 0.0,
            "title": "Phase Width Us",
            "type": "number"
          },
          "pitches_um": {
            "items": {
              "type": "number"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Pitches Um",
            "type": "array"
          },
          "sigma_S_per_m": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Sigma S Per M",
            "type": "number"
          },
          "trajectory_jitter_deg": {
            "default": 15.0,
            "exclusiveMinimum": 0.0,
            "maximum": 90.0,
            "title": "Trajectory Jitter Deg",
            "type": "number"
          },
          "trajectory_k": {
            "default": 1,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Trajectory K",
            "type": "integer"
          }
        },
        "title": "StudyControls",
        "type": "object"
      },
      "StudyPoint": {
        "description": "One evaluated geometry on the selectivity-versus-cost plane.",
        "properties": {
          "cost_uA": {
            "title": "Cost Ua",
            "type": "number"
          },
          "diameter_um": {
            "title": "Diameter Um",
            "type": "number"
          },
          "on_frontier": {
            "title": "On Frontier",
            "type": "boolean"
          },
          "pitch_um": {
            "title": "Pitch Um",
            "type": "number"
          },
          "safe": {
            "title": "Safe",
            "type": "boolean"
          },
          "selectivity_uA": {
            "title": "Selectivity Ua",
            "type": "number"
          },
          "spread_uA": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spread Ua"
          }
        },
        "required": [
          "diameter_um",
          "pitch_um",
          "cost_uA",
          "selectivity_uA",
          "safe",
          "on_frontier"
        ],
        "title": "StudyPoint",
        "type": "object"
      },
      "StudyResult": {
        "description": "A completed sweep: every activated geometry, and how many were swept.",
        "properties": {
          "n_geometries": {
            "title": "N Geometries",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/StudyPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "tier": {
            "enum": [
              "analytical",
              "fem"
            ],
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "points",
          "n_geometries",
          "tier"
        ],
        "title": "StudyResult",
        "type": "object"
      },
      "SweepControls": {
        "description": "An amplitude sweep of the scene: who fires, at what current.",
        "properties": {
          "amp_max_uA": {
            "default": 200.0,
            "exclusiveMinimum": 0.0,
            "title": "Amp Max Ua",
            "type": "number"
          },
          "amp_min_uA": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Amp Min Ua",
            "type": "number"
          },
          "body": {
            "discriminator": {
              "mapping": {
                "cad": "#/components/schemas/CadBodySpec",
                "cylinder": "#/components/schemas/CylinderBody",
                "frustum": "#/components/schemas/FrustumBody",
                "hemisphere": "#/components/schemas/HemisphereBody",
                "none": "#/components/schemas/NoBody"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NoBody"
              },
              {
                "$ref": "#/components/schemas/HemisphereBody"
              },
              {
                "$ref": "#/components/schemas/CylinderBody"
              },
              {
                "$ref": "#/components/schemas/FrustumBody"
              },
              {
                "$ref": "#/components/schemas/CadBodySpec"
              }
            ],
            "title": "Body"
          },
          "electrode_um": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Electrode Um",
            "type": "number"
          },
          "extent_um": {
            "default": 130.0,
            "exclusiveMinimum": 0.0,
            "title": "Extent Um",
            "type": "number"
          },
          "include_scorecard": {
            "default": false,
            "title": "Include Scorecard",
            "type": "boolean"
          },
          "layout": {
            "default": "single",
            "enum": [
              "single",
              "bipolar"
            ],
            "title": "Layout",
            "type": "string"
          },
          "n": {
            "default": 61,
            "maximum": 257.0,
            "minimum": 9.0,
            "title": "N",
            "type": "integer"
          },
          "n_amplitudes": {
            "default": 24,
            "maximum": 60.0,
            "minimum": 2.0,
            "title": "N Amplitudes",
            "type": "integer"
          },
          "neighbor_um": {
            "default": 40.0,
            "exclusiveMinimum": 0.0,
            "title": "Neighbor Um",
            "type": "number"
          },
          "overlap_policy": {
            "default": "reject",
            "enum": [
              "reject",
              "displace"
            ],
            "title": "Overlap Policy",
            "type": "string"
          },
          "phase_width_us": {
            "default": 200.0,
            "exclusiveMinimum": 0.0,
            "title": "Phase Width Us",
            "type": "number"
          },
          "pitch_um": {
            "default": 60.0,
            "exclusiveMinimum": 0.0,
            "title": "Pitch Um",
            "type": "number"
          },
          "sigma_S_per_m": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Sigma S Per M",
            "type": "number"
          },
          "spacing": {
            "default": "linear",
            "enum": [
              "linear",
              "log"
            ],
            "title": "Spacing",
            "type": "string"
          }
        },
        "title": "SweepControls",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Retinode API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/score": {
      "post": {
        "operationId": "submit_score_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneControls"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Score"
      }
    },
    "/study": {
      "post": {
        "operationId": "submit_study_study_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyControls"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Study"
      }
    },
    "/sweep": {
      "post": {
        "operationId": "submit_sweep_sweep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SweepControls"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Sweep"
      }
    },
...
  }
}